+++ packages/astro/src/assets/types.ts
@@ -11,7 +11,7 @@ export type ImageOutputFormat = (typeof VALID_OUTPUT_FORMATS)[number] | (string 
 export type AssetsGlobalStaticImagesList = Map<
 	string,
 	{
-		originalSrcPath: string;
+		originalSrcPath: string | undefined;
 		transforms: Map<string, { finalPath: string; transform: ImageTransform }>;
 	}
 >;
@@ -21,7 +21,7 @@ declare global {
 	var astroAsset: {
 		imageService?: ImageService;
 		addStaticImage?:
-			| ((options: ImageTransform, hashProperties: string[], fsPath: string) => string)
+			| ((options: ImageTransform, hashProperties: string[], fsPath: string | undefined) => string)
 			| undefined;
 		staticImages?: AssetsGlobalStaticImagesList;
 		referencedImages?: Set<string>;
@@ -118,6 +118,28 @@ type ImageSharedProps<T> = T & {
 	 * ```
 	 */
 	height?: number | `${number}`;
+	/**
+	 * Desired output format for the image. Defaults to `webp`.
+	 *
+	 * **Example**:
+	 * ```astro
+	 * <Image src={...} format="avif" alt="..." />
+	 * ```
+	 */
+	format?: ImageOutputFormat;
+	/**
+	 * Desired quality for the image. Value can either be a preset such as `low` or `high`, or a numeric value from 0 to 100.
+	 *
+	 * The perceptual quality of the output image is service-specific.
+	 * For instance, a certain service might decide that `high` results in a very beautiful image, but another could choose for it to be at best passable.
+	 *
+	 * **Example**:
+	 * ```astro
+	 * <Image src={...} quality='high' alt="..." />
+	 * <Image src={...} quality={300} alt="..." />
+	 * ```
+	 */
+	quality?: ImageQuality;
 } & (
 		| {
 				/**
@@ -153,28 +175,6 @@ export type LocalImageProps<T> = ImageSharedProps<T> & {
 	 * ```
 	 */
 	src: ImageMetadata | Promise<{ default: ImageMetadata }>;
-	/**
-	 * Desired output format for the image. Defaults to `webp`.
-	 *
-	 * **Example**:
-	 * ```astro
-	 * <Image src={...} format="avif" alt="..." />
-	 * ```
-	 */
-	format?: ImageOutputFormat;
-	/**
-	 * Desired quality for the image. Value can either be a preset such as `low` or `high`, or a numeric value from 0 to 100.
-	 *
-	 * The perceptual quality of the output image is service-specific.
-	 * For instance, a certain service might decide that `high` results in a very beautiful image, but another could choose for it to be at best passable.
-	 *
-	 * **Example**:
-	 * ```astro
-	 * <Image src={...} quality='high' alt="..." />
-	 * <Image src={...} quality={300} alt="..." />
-	 * ```
-	 */
-	quality?: ImageQuality;
 };
 
 export type RemoteImageProps<T> =