+++ packages/integrations/vercel/test/isr.test.js
@@ -34,8 +34,16 @@ describe('ISR', () => {
 				dest: '_render',
 			},
 			{
+				src: '^/excluded/([^/]+?)$',
+				dest: '_render',
+			},
+			{
 				src: '^\\/_image$',
 				dest: '_render',
+			},
+			{
+				src: '^\\/excluded\\/([^/]+?)\\/?$',
+				dest: '/_isr?x_astro_path=$0',
 			},
 			{
 				src: '^\\/one\\/?$',